+++ src/app/components/Icons/FavoritesIcon.tsx
@@ -1,6 +1,6 @@
-import SVG, { SVGProps } from '../SVG/SVG';
+import SVG, { SVGProps } from '../IconSvg/IconSvg';
 
-function IconFavorites(props: Omit<SVGProps, 'children'>): JSX.Element {
+function FavoritesIcon(props: Omit<SVGProps, 'children'>): JSX.Element {
   return (
     <SVG {...props}>
       <>
@@ -16,4 +16,4 @@ function IconFavorites(props: Omit<SVGProps, 'children'>): JSX.Element {
   );
 }
 
-export default IconFavorites;
+export default FavoritesIcon;
