+++ src/app/components/Icons/SettingsIcon.tsx
@@ -1,6 +1,6 @@
-import SVG, { SVGProps } from '../SVG/SVG';
+import SVG, { SVGProps } from '../IconSvg/IconSvg';
 
-function IconSettings(props: Omit<SVGProps, 'children'>): JSX.Element {
+function SettingsIcon(props: Omit<SVGProps, 'children'>): JSX.Element {
   return (
     <SVG {...props}>
       <>
@@ -11,4 +11,4 @@ function IconSettings(props: Omit<SVGProps, 'children'>): JSX.Element {
   );
 }
 
-export default IconSettings;
+export default SettingsIcon;
