+++ src/app/components/Sidebar/Sidebar.tsx
@@ -1,24 +1,24 @@
 import { classNames } from '../../utils/styles';
 import classes from './Sidebar.module.css';
-import { Settings, Info, Home, Favorites } from '../Icons';
+import { SettingsIcon, InfoIcon, HomeIcon, FavoritesIcon } from '../Icons';
 
 function Sidebar(): JSX.Element {
   return (
     <nav className={classNames(classes.container, classes.flex)}>
       <section className={classNames(classes.section, classes.flex)}>
         <a className={classes.selected}>
-          <Home selected={true} />
+          <HomeIcon selected={true} />
         </a>
         <a className={classes.disabled} data-tooltip="coming soon">
-          <Favorites disabled={true} />
+          <FavoritesIcon disabled={true} />
         </a>
       </section>
       <aside className={classes.flex}>
         <a className={classes.disabled} data-tooltip="coming soon">
-          <Info disabled={true} />
+          <InfoIcon disabled={true} />
         </a>
         <a className={classes.disabled} data-tooltip="coming soon">
-          <Settings disabled={true} />
+          <SettingsIcon disabled={true} />
         </a>
       </aside>
     </nav>
